refactor: unify CreateIconConfig and IconGlyphSource into one type
An icon set was described by two identical types living in two modules,
which also forced IconGlyphSourceContext to re-export a type it did not
own. Collapse them into IconGlyphSource, defined alongside its sibling
types in createIcon.
