+++ packages/mobile/src/icons/index.ts
@@ -1,7 +1,7 @@
-export type { CreateIconConfig, GlyphMap, IconGlyphResolverArgs } from './createIcon';
+export type { GlyphMap, IconGlyphResolverArgs, IconGlyphSource } from './createIcon';
 export { createIcon, DEFAULT_ICON_FONT_FAMILY } from './createIcon';
 export * from './Icon';
-export type { IconGlyphSource, IconGlyphSourceProviderProps } from './IconGlyphSourceContext';
+export type { IconGlyphSourceProviderProps } from './IconGlyphSourceContext';
 export { IconGlyphSourceProvider, useIconGlyphSources } from './IconGlyphSourceContext';
 export * from './LogoMark';
 export * from './LogoWordmark';
