+++ packages/web/src/icons/index.ts
@@ -1,8 +1,8 @@
 export * from './Icon';
-export type { IconGlyphSource, IconGlyphSourceProviderProps } from './IconGlyphSourceContext';
+export type { IconGlyphSourceProviderProps } from './IconGlyphSourceContext';
 export { IconGlyphSourceProvider, useIconGlyphSources } from './IconGlyphSourceContext';
 export { createIcon, DEFAULT_ICON_FONT_FAMILY } from './createIcon';
-export type { CreateIconConfig, GlyphMap, IconGlyphResolverArgs } from './createIcon';
+export type { GlyphMap, IconGlyphResolverArgs, IconGlyphSource } from './createIcon';
 export * from './LogoMark';
 export * from './LogoWordmark';
 export * from './SubBrandLogoMark';
